Completed implementation of DMTK Window Resizers.
Resizers allow the user to adjust the size of their DMTK Windows. The resizers appear in the corners of the window when hovered over and can be clicked and dragged to adjust.

Going to look into the de-synced movements of the mouse when resizing windows. I assume I need to factor in resolution somewhere?