Solution to resizer dragging bug.
This one took a good while to find. All UI is parented under a Canvas. This canvas usually has a Canvas Scalar component attached to it which changes a scalarFactor value within the Canvas. To ensure the scaling occurs correctly, the mouse positions need to be converted into canvas space by dividing them by the scalarFactor value.